talklens/requirements.txt
# ─── Core Framework ───────────────────────────────────────────────────────────
# Python 3.9 compatible — tested on macOS arm64 (Apple Silicon) & x86_64
streamlit>=1.32.0,<1.36.0
streamlit-webrtc>=0.47.0,<0.48.0

# ─── Computer Vision ──────────────────────────────────────────────────────────
opencv-python>=4.9.0.80,<4.11.0
ultralytics>=8.2.0,<8.3.0           # YOLOv8
mediapipe>=0.10.30                   # latest wheel available for Python 3.9

# ─── Deep Learning ────────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0,<2.0.0               # numpy 2.x breaks several deps on 3.9
scikit-learn>=1.3.0,<1.6.0

# ─── Speech Processing ────────────────────────────────────────────────────────
openai-whisper>=20231117
pyttsx3>=2.90
SpeechRecognition>=3.10.0
# pyaudio — install separately via brew if needed:
#   brew install portaudio && pip install pyaudio
# Uncomment line below only after portaudio is installed:
# pyaudio>=0.2.14

# ─── NLP / Language ───────────────────────────────────────────────────────────
transformers>=4.40.0,<4.43.0
sentence-transformers>=2.7.0,<3.1.0

# ─── Data & Utilities ─────────────────────────────────────────────────────────
pandas>=2.0.0,<2.3.0
Pillow>=10.0.0,<11.0.0
tqdm>=4.65.0
pydantic>=2.5.0,<2.8.0
python-dotenv>=1.0.0
loguru>=0.7.0
av==13.1.0                          # pre-built wheel for cp39 arm64; >=14.x requires building from source

# ─── Optional: Groq/OpenAI/Gemini LLM integration ─────────────────────────────
groq>=0.9.0
openai>=1.30.0,<2.0.0
google-generativeai>=0.5.0

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-mock>=3.12.0
